requirements.txt
streamlit
fastapi
uvicorn
llama-index
llama-index-llms-ollama
faiss-cpu
sentence-transformers
arxiv
transformers
torch
plotly
pandas
rank_bm25
requests
